Refactor: вынес функцию загрузки сообщений в отдельный модуль
